spaces_upload/requirements.txt
gradio
pillow
torch
torchvision
numpy
